pyproject.toml
[project]
name = "strava-dashboard"
version = "0.1.0"
description = "Interactive Strava Data Dashboard"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "plotly>=5.17.0",
    "requests>=2.31.0",
    "folium>=0.15.0",
    "python-dotenv>=1.0.0",
    "stravalib>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
